requirements.txt
Django==2.2.2
djangorestframework==3.11.0
django-cors-headers==3.2.0